Moved methods from readdata.py to jafem.py and removed readdata.py
